Console Formatting!
changed deque to vector. makes it simpler and easier to work with
Fixed compilation warnings
Apparently I can't use "ERROR" as an enum key...
Moved Logging functions into a .cpp
turned back const char*& to const char*

+++ Utility/src/Composing/Formattables/FInteger.cpp
@@ -63,7 +63,7 @@ namespace Composing::Formattables
       {
          std::string x_Padded;
          x_Padded.reserve( x_Padding );
-         int x_CharDiff = x_Padding - x_Out.length();
+         long long x_CharDiff = x_Padding - x_Out.length();
 
          if ( x_CharDiff > 0 )
          {
